+++ lua/plugins/lsp.lua
@@ -126,10 +126,10 @@ local function configure_eslint(lspconfig, capabilities)
   lspconfig.eslint.setup({
     capabilities = capabilities,
     on_attach = function(_, bufnr)
-      vim.api.nvim_create_autocmd("BufWritePre", {
-        buffer = bufnr,
-        command = "EslintFixAll",
-      })
+      -- vim.api.nvim_create_autocmd("BufWritePre", {
+      --   buffer = bufnr,
+      --   command = "EslintFixAll",
+      -- })
       on_attach(bufnr)
     end,
   })